# openapi_with_servers.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chip's AI Operating Partner API",
    "description": "KPI trends, email search, and intelligent Q&A for Chip Ridge's portfolio.",
    "version": "0.1.0"
  },
  "servers": [
    { "url": "https://ai-ops-internal-production.up.railway.app" }
  ],
  "paths": {
    "/ask": {
      "post": {
        "tags": ["Q&A"],
        "summary": "Ask",
        "description": "Answer a natural-language question using KPI Sheet + Email RAG.",
        "operationId": "ask_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/kpis": {
      "get": {
        "tags": ["KPI Sheet"],
        "summary": "Get Kpis",
        "description": "Return all rows from the DAILY_KPI_SNAPSHOT Google Sheet.",
        "operationId": "get_kpis_kpis_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KPIResponse"
                }
              }
            }
          }
        }
      }
    },
    "/search-emails": {
      "post": {
        "tags": ["Email RAG"],
        "summary": "Search Emails",
        "description": "Semantic search over indexed emails (ChromaDB).\n\nSupports optional metadata filters: folder, sender_domain, date range.",
        "operationId": "search_emails_search_emails_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": ["System"],
        "summary": "Health",
        "description": "Health check — reports index size, sheet connectivity, key status.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/upload-emails": {
      "post": {
        "tags": ["Admin"],
        "summary": "Upload Emails",
        "description": "Bulk upsert email documents into the Railway ChromaDB index. Accepts pre-chunked email text + metadata in batches.",
        "operationId": "upload_emails_admin_upload_emails_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadEmailsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadEmailsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AskRequest": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question",
            "description": "Natural language question from Chip"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "LLM model for synthesis",
            "default": "gpt-4o"
          },
          "n_results": {
            "type": "integer",
            "maximum": 50,
            "minimum": 1,
            "title": "N Results",
            "description": "Max email hits for RAG path",
            "default": 10
          }
        },
        "type": "object",
        "required": ["question"],
        "title": "AskRequest"
      },
      "AskResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "sources": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Sources"
          },
          "paths_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Paths Used"
          },
          "cost_estimate_usd": {
            "type": "number",
            "title": "Cost Estimate Usd"
          },
          "tokens": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens"
          },
          "rag_debug": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rag Debug"
          }
        },
        "type": "object",
        "required": ["answer", "sources", "paths_used", "cost_estimate_usd"],
        "title": "AskResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "email_index_count": {
            "type": "integer",
            "title": "Email Index Count"
          },
          "kpi_sheet_rows": {
            "type": "integer",
            "title": "Kpi Sheet Rows"
          },
          "openai_key_set": {
            "type": "boolean",
            "title": "Openai Key Set"
          },
          "chromadb_path": {
            "type": "string",
            "title": "Chromadb Path",
            "default": ""
          },
          "sheet_connected": {
            "type": "boolean",
            "title": "Sheet Connected",
            "default": false
          }
        },
        "type": "object",
        "required": ["status", "email_index_count", "kpi_sheet_rows", "openai_key_set"],
        "title": "HealthResponse"
      },
      "KPIResponse": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/KPIRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": ["rows", "total"],
        "title": "KPIResponse"
      },
      "KPIRow": {
        "properties": {
          "row_number": {
            "type": "integer",
            "title": "Row Number"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": ["row_number", "data"],
        "title": "KPIRow"
      },
      "SearchHit": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Document chunk ID",
            "default": ""
          },
          "subject": {
            "type": "string",
            "title": "Subject",
            "default": ""
          },
          "sender": {
            "type": "string",
            "title": "Sender",
            "default": ""
          },
          "date": {
            "type": "string",
            "title": "Date",
            "default": ""
          },
          "folder": {
            "type": "string",
            "title": "Folder",
            "default": ""
          },
          "snippet": {
            "type": "string",
            "title": "Snippet",
            "description": "First ~500 chars of the matched document",
            "default": ""
          },
          "distance": {
            "anyOf": [
              {"type": "number"},
              {"type": "null"}
            ],
            "title": "Distance"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "title": "SearchHit"
      },
      "SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "Semantic search query"
          },
          "n_results": {
            "type": "integer",
            "maximum": 50,
            "minimum": 1,
            "title": "N Results",
            "default": 10
          },
          "folder": {
            "anyOf": [
              {"type": "string"},
              {"type": "null"}
            ],
            "title": "Folder",
            "description": "Filter by folder name (e.g. 'Inbox', 'Sent Items')"
          },
          "sender_domain": {
            "anyOf": [
              {"type": "string"},
              {"type": "null"}
            ],
            "title": "Sender Domain",
            "description": "Filter by sender domain (e.g. 'perpetualtitle.com')"
          },
          "date_from": {
            "anyOf": [
              {"type": "string"},
              {"type": "null"}
            ],
            "title": "Date From",
            "description": "Filter emails on or after this date (YYYY-MM-DD)"
          },
          "date_to": {
            "anyOf": [
              {"type": "string"},
              {"type": "null"}
            ],
            "title": "Date To",
            "description": "Filter emails on or before this date (YYYY-MM-DD)"
          }
        },
        "type": "object",
        "required": ["query"],
        "title": "SearchRequest"
      },
      "SearchResponse": {
        "properties": {
          "hits": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "type": "array",
            "title": "Hits"
          },
          "total_indexed": {
            "type": "integer",
            "title": "Total Indexed"
          }
        },
        "type": "object",
        "required": ["hits", "total_indexed"],
        "title": "SearchResponse"
      },
      "UploadEmailDoc": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Stable unique document ID"
          },
          "document": {
            "type": "string",
            "title": "Document",
            "description": "Text content of the email chunk"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Email metadata (subject, sender, date, folder, ...)"
          }
        },
        "type": "object",
        "required": ["id", "document"],
        "title": "UploadEmailDoc"
      },
      "UploadEmailsRequest": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/UploadEmailDoc"
            },
            "type": "array",
            "title": "Documents",
            "description": "List of documents to upsert"
          }
        },
        "type": "object",
        "required": ["documents"],
        "title": "UploadEmailsRequest"
      },
      "UploadEmailsResponse": {
        "properties": {
          "upserted": {
            "type": "integer",
            "title": "Upserted"
          },
          "total_in_collection": {
            "type": "integer",
            "title": "Total In Collection"
          }
        },
        "type": "object",
        "required": ["upserted", "total_in_collection"],
        "title": "UploadEmailsResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {"type": "string"},
                {"type": "integer"}
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}